ハンドラテストをエンドポイント単位のファイルに分割
announcement_test.go に集約していたテストを、対応するハンドラ
ファイル名に揃えて get/update/delete/list の4ファイルへ分割し、
テストの所在を見つけやすくした。テスト内容自体は変更していない。
